Update the common version referenced in the manifest files.
This is not currently used or verified, however it will become important once these manifest files move to downstream SDK repositories.
In the short term it still has utility in indicating clearly that these manifests were produced against the Alpha 1 state of the canonical features list, for which I shall push a Git tag shortly.