Sort the SDK columns alphabetically.
This guarantees consistency from render to render, expressing that intention in the code, rather than relying on natural ordering.

Also, previously, we were relying on Map's implementation following the same order with its forEach() method that it returns in keys(). It looks, from MDN docs, like Map does state "in insertion order" for both forEach() and keys() - however that feels like a subtle implementation detail upon which we should not rely for something so important. Hence the change to something more explicit.